+++ flatarray/shared/src/main/scala/me/kright/gametools/flatarray/FlatDoubleSerializer.scala
@@ -27,59 +27,68 @@ object FlatDoubleSerializer:
     override def read(array: Array[Double], offset: Int): T =
       FlatDoubleSerializer.read[T](array, offset)
 
-  
-  //* count of double values in case class */
+
+  //* count of double values in case class, recursing into nested case classes */
   inline def getSize[T]: Int = ${ getSizeImpl[T] }
 
   private def getSizeImpl[T: Type](using Quotes): Expr[Int] =
     import quotes.reflect.*
 
     val tpe = TypeRepr.of[T]
-    val symbol = tpe.typeSymbol
 
-    if (!symbol.flags.is(Flags.Case)) {
+    if (!tpe.typeSymbol.flags.is(Flags.Case)) {
       report.errorAndAbort(s"Type ${Type.show[T]} must be a case class")
     }
 
-    val fields = tpe.typeSymbol.caseFields
-    val doubleFields = fields.filter(field => TypeRepr.of[Double] =:= tpe.memberType(field))
+    Expr(countDoubles(tpe))
 
-    if (doubleFields.length != fields.length) {
-      report.errorAndAbort(s"All fields in ${Type.show[T]} must be of type Double")
-    }
+  private def countDoubles(using q: Quotes)(tpe: q.reflect.TypeRepr): Int =
+    import q.reflect.*
 
-    Expr(fields.length)
+    if (tpe =:= TypeRepr.of[Double])
+      1
+    else if (tpe.typeSymbol.flags.is(Flags.Case))
+      tpe.typeSymbol.caseFields.map(field => countDoubles(tpe.memberType(field))).sum
+    else
+      report.errorAndAbort(s"Field type ${tpe.show} must be Double or a case class with Double fields")
 
-  /** write case class with double fields into the array of doubles at specific offset */
+  /** write case class with double fields (or nested case classes of doubles) into the array of doubles at specific offset */
   inline def write[T](elem: T, array: Array[Double], offset: Int): Unit = ${ writeImpl[T]('elem, 'array, 'offset) }
 
   private def writeImpl[T: Type](elem: Expr[T], array: Expr[Array[Double]], offset: Expr[Int])(using Quotes): Expr[Unit] =
     import quotes.reflect.*
 
     val tpe = TypeRepr.of[T]
-    val symbol = tpe.typeSymbol
 
-    if (!symbol.flags.is(Flags.Case)) {
+    if (!tpe.typeSymbol.flags.is(Flags.Case)) {
       report.errorAndAbort(s"Type ${Type.show[T]} must be a case class")
     }
 
-    val fields = tpe.typeSymbol.caseFields
-    val doubleFields = fields.filter(field => TypeRepr.of[Double] =:= tpe.memberType(field))
+    val (assignments, _) = collectWrites(elem.asTerm, tpe, 0, array, offset)
 
-    if (doubleFields.length != fields.length) {
-      report.errorAndAbort(s"All fields in ${Type.show[T]} must be of type Double")
-    }
-
-    val assignments = fields.zipWithIndex.map { (field, idx) =>
-      val fieldSelect = Select(elem.asTerm, field).asExprOf[Double]
-      val arrayIdx = Expr(idx)
-      '{ $array($offset + $arrayIdx) = $fieldSelect }
-    }
-
-    val result = assignments.foldLeft[Expr[Unit]]('{ () })((acc, expr) =>
+    assignments.foldLeft[Expr[Unit]]('{ () })((acc, expr) =>
       '{ $acc ; $expr }
     )
-    result
+
+  private def collectWrites(using q: Quotes)(term: q.reflect.Term,
+                                             tpe: q.reflect.TypeRepr,
+                                             startIdx: Int,
+                                             array: Expr[Array[Double]],
+                                             offset: Expr[Int]): (List[Expr[Unit]], Int) =
+    import q.reflect.*
+
+    if (tpe =:= TypeRepr.of[Double]) {
+      val arrayIdx = Expr(startIdx)
+      val fieldValue = term.asExprOf[Double]
+      (List('{ $array($offset + $arrayIdx) = $fieldValue }), startIdx + 1)
+    } else if (tpe.typeSymbol.flags.is(Flags.Case)) {
+      tpe.typeSymbol.caseFields.foldLeft((List.empty[Expr[Unit]], startIdx)) { case ((acc, idx), field) =>
+        val (writes, nextIdx) = collectWrites(Select(term, field), tpe.memberType(field), idx, array, offset)
+        (acc ++ writes, nextIdx)
+      }
+    } else {
+      report.errorAndAbort(s"Field type ${tpe.show} must be Double or a case class with Double fields")
+    }
 
   /** restore case class back from the array at specific offset */
   inline def read[T](array: Array[Double], offset: Int): T = ${ readImpl[T]('array, 'offset) }
@@ -88,25 +97,30 @@ object FlatDoubleSerializer:
     import quotes.reflect.*
 
     val tpe = TypeRepr.of[T]
-    val symbol = tpe.typeSymbol
 
-    if (!symbol.flags.is(Flags.Case)) {
+    if (!tpe.typeSymbol.flags.is(Flags.Case)) {
       report.errorAndAbort(s"Type ${Type.show[T]} must be a case class")
     }
 
-    val fields = tpe.typeSymbol.caseFields
-    val doubleFields = fields.filter(field => TypeRepr.of[Double] =:= tpe.memberType(field))
+    val (result, _) = buildRead(tpe, 0, array, offset)
+    result.asExprOf[T]
 
-    if (doubleFields.length != fields.length) {
-      report.errorAndAbort(s"All fields in ${Type.show[T]} must be of type Double")
+  private def buildRead(using q: Quotes)(tpe: q.reflect.TypeRepr,
+                                         startIdx: Int,
+                                         array: Expr[Array[Double]],
+                                         offset: Expr[Int]): (q.reflect.Term, Int) =
+    import q.reflect.*
+
+    if (tpe =:= TypeRepr.of[Double]) {
+      val arrayIdx = Expr(startIdx)
+      ('{ $array($offset + $arrayIdx) }.asTerm, startIdx + 1)
+    } else if (tpe.typeSymbol.flags.is(Flags.Case)) {
+      val symbol = tpe.typeSymbol
+      val (args, nextIdx) = symbol.caseFields.foldLeft((List.empty[Term], startIdx)) { case ((acc, idx), field) =>
+        val (arg, next) = buildRead(tpe.memberType(field), idx, array, offset)
+        (acc :+ arg, next)
+      }
+      (Apply(Select(New(Inferred(tpe)), symbol.primaryConstructor), args), nextIdx)
+    } else {
+      report.errorAndAbort(s"Field type ${tpe.show} must be Double or a case class with Double fields")
     }
-
-    val fieldValues = fields.zipWithIndex.map { (field, idx) =>
-      val arrayIdx = Expr(idx)
-      '{ $array($offset + $arrayIdx) }
-    }
-
-    val newExpr = New(TypeTree.of[T])
-    val constructor = symbol.primaryConstructor
-    val newInstance = Apply(Select(newExpr, constructor), fieldValues.map(_.asTerm)).asExprOf[T]
-    newInstance