+++ pga2dgeom/shared/src/main/scala/me/kright/gametools/pga2d/geom/Pga2dEdge.scala
@@ -1,11 +1,12 @@
 package me.kright.gametools.pga2d.geom
 
+import me.kright.gametools.flatarray.FlatDoubleSerializer
 import me.kright.gametools.mathutil.MathUtil.clamp
 import me.kright.gametools.pga2d.*
 import me.kright.gametools.mathutil.MathUtil
 
 case class Pga2dEdge(a: Pga2dPoint,
-                     b: Pga2dPoint) derives CanEqual:
+                     b: Pga2dPoint) derives CanEqual, FlatDoubleSerializer:
 
   override def toString: String = s"Pga2dEdge(a = $a, b = $b)"