+++ pga3dgeom/shared/src/main/scala/me/kright/gametools/pga3d/geom/Pga3dCylinder.scala
@@ -1,5 +1,6 @@
 package me.kright.gametools.pga3d.geom
 
+import me.kright.gametools.flatarray.FlatDoubleSerializer
 import me.kright.gametools.pga3d.{Pga3dBivector, Pga3dPoint, Pga3dVector}
 
 /**
@@ -10,7 +11,7 @@ import me.kright.gametools.pga3d.{Pga3dBivector, Pga3dPoint, Pga3dVector}
  */
 case class Pga3dCylinder(a: Pga3dPoint,
                          b: Pga3dPoint,
-                         r: Double):
+                         r: Double) derives CanEqual, FlatDoubleSerializer:
 
   def ab: Pga3dVector =
     b - a